+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLInputPeerUser.java
@@ -6,9 +6,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
 import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
-import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
@@ -22,7 +20,7 @@ public class TLInputPeerUser extends TLAbsInputPeer {
 
     public static final int CONSTRUCTOR_ID = 0x7b8e7de6;
 
-    protected int userId;
+    protected long userId;
 
     protected long accessHash;
 
@@ -31,21 +29,21 @@ public class TLInputPeerUser extends TLAbsInputPeer {
     public TLInputPeerUser() {
     }
 
-    public TLInputPeerUser(int userId, long accessHash) {
+    public TLInputPeerUser(long userId, long accessHash) {
         this.userId = userId;
         this.accessHash = accessHash;
     }
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(userId, stream);
+        writeLong(userId, stream);
         writeLong(accessHash, stream);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        userId = readInt(stream);
+        userId = readLong(stream);
         accessHash = readLong(stream);
     }
 
@@ -67,11 +65,11 @@ public class TLInputPeerUser extends TLAbsInputPeer {
         return CONSTRUCTOR_ID;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
